requirements.txt
altgraph==0.17.4
attrs==23.2.0
certifi==2024.2.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
h11==0.14.0
idna==3.7
lxml==5.2.1
numpy==1.26.4
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wsproto==1.2.0
